requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
beautifulsoup4==4.12.3
Faker==25.1.0
lxml==5.2.2
playwright==1.43.0 # Версия 1.43.0 не встала, необходимо разобраться
pyTelegramBotAPI==4.17.0
pytest-dependency==0.6.0
pytest-playwright==0.5.0
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
pytest==8.2.0
python-dotenv==1.0.1
requests==2.31.0
ruff==0.4.4
scipy==1.13.0